Implements piping from and to ffmpeg
While upscaling, the frames will be piped to the output folder in
batches, upscaled and then piped to ffmpeg to generate the output video.
Both ffmpeg processes run throughout the upscaling process to keep
timings accurate. This greatly reduces the hard drive space usage since
all the original and upscaled frames don't need to be stored.